Fixing the request_body as key issue
This is to fix https://github.com/zoltan-fedor/falcon-caching/issues/3
It turns out that by reading the request body from the `req.stream`, we are fully consuming it and so it is no longer available for the responder.

For now I just exclude the use of the request body in the cache key - if a feature request comes up later to include it, then will consider how we might be able to use it while still leaving it available for the responder.

Signed-off-by: Zoltan Fedor <zoltan.1.fedor@gmail.com>